# server/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ad-server API",
    "version": "1.0.0",
    "description": "Multi-tenant ad-serving API for digital signage / DOOH inventory. Generated at build time from this repo's Zod request schemas — see PROJECT_PLAN.md's 04j section for how and why."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this spec"
    }
  ],
...
  "paths": {
...
    "/api/admin/tenants": {
      "get": {
        "tags": [
          "Superadmin — Tenants"
        ],
        "summary": "List all tenants",
        "description": "**Auth:** superadmin",
        "security": [
          {
            "superadminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "tenants": []
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Superadmin — Tenants"
        ],
        "summary": "Create a tenant (atomically creates first tenant_admin invite)",
        "description": "**Auth:** superadmin",
        "security": [
          {
            "superadminAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "fulfillment_quota": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "admin_email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": [
                  "name",
                  "fulfillment_quota",
                  "admin_email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "tenant": {},
                  "invite": {
                    "invite_url": "https://...",
                    "expires_at": "2026-07-19T00:00:00.000Z"
                  },
                  "api_key": "plaintext-shown-once"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/tenants/{id}": {
      "patch": {
        "tags": [
          "Superadmin — Tenants"
        ],
        "summary": "Update tenant status/quota/timeout",
        "description": "**Auth:** superadmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "superadminAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "deactivated"
                    ]
                  },
                  "fulfillment_quota": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "reservation_timeout_seconds": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "tenant": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Superadmin — Tenants"
        ],
        "summary": "Get a tenant with its campaigns and screens",
        "description": "**Auth:** superadmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "superadminAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "tenant": {},
                  "campaigns": [],
                  "screens": []
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tenant/api-key": {
      "get": {
        "tags": [
          "Tenant — API key & usage"
        ],
        "summary": "View tenant API key status",
        "description": "**Auth:** tenant_admin JWT or tenant API key",
        "security": [
          {
            "tenantAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "key_prefix": "abcd1234",
                  "status": "active"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/api-key/rotate": {
      "post": {
        "tags": [
          "Tenant — API key & usage"
        ],
        "summary": "Rotate tenant API key",
        "description": "**Auth:** tenant_admin JWT or tenant API key",
        "security": [
          {
            "tenantAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "api_key": "plaintext-shown-once"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/usage": {
      "get": {
        "tags": [
          "Tenant — API key & usage"
        ],
        "summary": "Tenant-wide fulfillment quota usage",
        "description": "**Auth:** tenant_admin JWT or tenant API key",
        "security": [
          {
            "tenantAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "used": 0,
                  "quota": 1000
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/usage/by-screen": {
      "get": {
        "tags": [
          "Tenant — API key & usage"
        ],
        "summary": "Windowed fulfillment usage broken down per screen",
        "description": "**Auth:** tenant_admin JWT or tenant API key",
        "security": [
          {
            "tenantAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "window_hours": 24,
                  "screens": []
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/play-log": {
      "get": {
        "tags": [
          "Tenant — Play log"
        ],
        "summary": "Cursor-paginated log of media plays (fulfillments) with campaign/screen info",
        "description": "**Auth:** tenant_admin JWT or tenant API key",
        "security": [
          {
            "tenantAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "entries": [],
                  "next_cursor": null
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/play-log/export": {
      "get": {
        "tags": [
          "Tenant — Play log"
        ],
        "summary": "CSV download of the play log for the past day/week/month",
        "description": "**Auth:** tenant_admin JWT or tenant API key",
        "security": [
          {
            "tenantAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": "text/csv attachment"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "superadminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Supabase Auth JWT for a superadmin user"
      },
      "tenantAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "tenant_admin JWT or tenant API key — server/tenant-access.ts accepts either"
      },
      "deviceAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Device API key issued at screen registration (POST /v1/screens)"
      },
      "cronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET — only enforced when that env var is set"
      }
    }
  }
}